refact: MemberViewController를 분리된 화면 전용 컨트롤러 및 URL에 등록 API 명시
