+++ src/main/resources/templates/member/list.html
@@ -27,7 +27,7 @@
     </tbody>
 </table>
 </body>
-<form th:action="@{/logout}" method="post">
+<form th:action="@{/admin/members/logout}" method="post">
     <button type="submit">로그아웃</button>
 </form>
 </html>
